mismatch_query_generator/requirements.txt
google-cloud-bigquery>=2.32.0
google-cloud-storage>=1.28.1
nltk>=3.7 